requirements.txt
selenium
undetected-chromedriver
curl-cffi
python-dotenv